+++ scripts/build_gitlab/osx.sh
@@ -22,7 +22,6 @@ rm -rf *
 # NOTE(jda) - using Internal tasking system here temporarily to avoid installing TBB
 cmake \
 -D OSPRAY_TASKING_SYSTEM=Internal \
--D OSPRAY_MODULE_BILINEAR_PATCH=ON \
 -D OSPRAY_ENABLE_TESTING=ON \
 -D OSPRAY_SG_CHOMBO=OFF \
 -D OSPRAY_SG_OPENIMAGEIO=OFF \